added entity, dto, service and repo

+++ src/main/java/net/codejava/repository/UserRepository.java
@@ -1,14 +1,14 @@
 package net.codejava.repository;
 
-import java.util.Optional;
-
 import net.codejava.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-	
-	Optional<User> findByEmail(String email);
-	
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+
 }
